// src/app/user/user.interface.ts
export type Order = {
    productName: string
    price: number
    quantity: number
  }
  
  export type UserInfo = {
    userId: number
    username: string
    password: string
    fullName: {
      firstName: string
      lastName: string
    }
    age: number
    email: string
    isActive: boolean
    hobbies: string[]
    address: {
      street: string
      city: string
      country: string
    }
    orders?: Order[]
  }
  